finishing orphans

+++ controllers/orphans.controller.js
@@ -9,18 +9,67 @@ const requestStatus = require("../utilities/requestStatus");
 
 // Get all orphans with pagination: ok
 const getAllOrphans = asyncWrapper(
-    async (req, res) => {
+    async (req, res, next) => {
         const query = req.query;
         
         const limit = query.limit || 4;
         const page = query.page || 1;
         const skip = (page - 1) * limit;
-        
+
         const orphans = await Orphan.find({}, { __v: false })
             .limit(limit)
             .skip(skip)
             .populate("orphanage", "name location")
             .populate("orphanageAdmin", "name email");
+
+        if(!orphans.length === 0){
+            const error = appError.create("No Orphans In Your Orphanage Yet!", 400, httpStatusText.FAIL);
+            return next(error);
+        }
+
+        return res.json({ status: httpStatusText.SUCCESS, data: { orphans } });
+    }
+);
+
+// Get all orphans of orphanage with pagination: ok
+const getAllOrphansOfOrphanage = asyncWrapper(
+    async (req, res, next) => {
+        const query = req.query;
+        
+        const limit = query.limit || 4;
+        const page = query.page || 1;
+        const skip = (page - 1) * limit;
+
+        const orphanageId = req.params.id;
+        const orphanageAdmin = req.currentUser.id;
+        const orphanageDoc = await Orphanage.findById(new mongoose.Types.ObjectId(orphanageId));
+    
+        if (!orphanageDoc) {
+            const error = appError.create("Orphanage not found", 400, httpStatusText.FAIL);
+            return next(error);
+        }
+
+        if (orphanageDoc.status !== requestStatus.APPROVED){
+            const error = appError.create("Orphanage not approved", 400, httpStatusText.FAIL);
+            return next(error);
+        }
+
+        if (!orphanageDoc.admin.equals(orphanageAdmin)) {
+            const error = appError.create("Orphanage Admin mismatch", 400, httpStatusText.FAIL);
+            return next(error);
+        }
+
+        // find orphan by it orphanage id:
+        const orphans = await Orphan.find({orphanage: orphanageDoc._id}, { __v: false })
+            .limit(limit)
+            .skip(skip)
+            .populate("orphanage", "name location")
+            .populate("orphanageAdmin", "name email");
+
+        if(!orphans.length === 0){
+            const error = appError.create("No Orphans In Your Orphanage Yet!", 400, httpStatusText.FAIL);
+            return next(error);
+        }
 
         return res.json({ status: httpStatusText.SUCCESS, data: { orphans } });
     }
@@ -30,12 +79,48 @@ const getAllOrphans = asyncWrapper(
 const getOrphanById = asyncWrapper(
     async (req, res, next) => {
         const id = req.params.id;
-        const orphan = await Orphan.findById(id)
+        const orphan = await Orphan.findById(new mongoose.Types.ObjectId(id))
             .populate("orphanage", "name location")
             .populate("orphanageAdmin", "name email");
 
         if (!orphan) {
             const error = appError.create("Orphan not found", 400, httpStatusText.FAIL);
+            return next(error);
+        }
+
+        return res.json({ status: httpStatusText.SUCCESS, data: { orphan } });
+    }
+);
+
+// Get orphan by ID in orphanage: ok
+const getOrphanByIdInOrphanage = asyncWrapper(
+    async (req, res, next) => {
+        const orphanageAdmin = req.currentUser.id;
+        const orphanageId = req.params.orphanageid;
+        const orphanageDoc = await Orphanage.findById(new mongoose.Types.ObjectId(orphanageId));
+    
+        if (!orphanageDoc) {
+            const error = appError.create("Orphanage not found", 400, httpStatusText.FAIL);
+            return next(error);
+        }
+
+        if (orphanageDoc.status !== requestStatus.APPROVED){
+            const error = appError.create("Orphanage not approved", 400, httpStatusText.FAIL);
+            return next(error);
+        }
+
+        if (!orphanageDoc.admin.equals(orphanageAdmin)) {
+            const error = appError.create("Orphanage Admin mismatch", 400, httpStatusText.FAIL);
+            return next(error);
+        }
+
+        const id = req.params.id;
+        const orphan = await Orphan.find({orphanage: orphanageDoc._id, _id: id})
+            .populate("orphanage", "name location")
+            .populate("orphanageAdmin", "name email");
+
+        if (orphan.length === 0) {
+            const error = appError.create("There is No Orphan With This ID In Your Orphanage!", 400, httpStatusText.FAIL);
             return next(error);
         }
 
@@ -134,12 +219,39 @@ const updateOrphan = asyncWrapper(async (req, res, next) => {
 // Delete orphan
 const deleteOrphan = asyncWrapper(async (req, res, next) => {
     const orphanId = req.params.id;
+    const orphanageAdmin = req.currentUser.id;
 
-    const orphan = await Orphan.findByIdAndDelete(orphanId);
+    const oldOrphan = await Orphan.findById(orphanId);
 
-    if (!orphan) {
+    if (!oldOrphan) {
         return next(appError.create("Orphan not found", 404, httpStatusText.FAIL));
     }
+    
+    const orphanageDoc = await Orphanage.findById(oldOrphan.orphanage);
+    
+    if (!orphanageDoc) {
+        const error = appError.create("Orphanage not found", 400, httpStatusText.FAIL);
+        return next(error);
+    }
+
+    if (orphanageDoc.status !== requestStatus.APPROVED){
+        const error = appError.create("Orphanage not approved", 400, httpStatusText.FAIL);
+        return next(error);
+    }
+
+    if (!orphanageDoc.admin.equals(orphanageAdmin)) {
+        const error = appError.create("Orphanage Admin mismatch", 400, httpStatusText.FAIL);
+        return next(error);
+    }
+
+        // Delete the orphan
+        await Orphan.findByIdAndDelete(orphanId);
+
+        // Remove orphan from orphanage's orphans array
+        orphanageDoc.orphans = orphanageDoc.orphans.filter(id => !id.equals(orphanId));
+        
+        // Save the updated orphanage document
+        await orphanageDoc.save();
 
     res.json({
         status: httpStatusText.SUCCESS,
@@ -149,7 +261,9 @@ const deleteOrphan = asyncWrapper(async (req, res, next) => {
 
 module.exports = {
     getAllOrphans,
+    getAllOrphansOfOrphanage,
     getOrphanById,
+    getOrphanByIdInOrphanage,
     createOrphan,
     updateOrphan,
     deleteOrphan
